// epc-frontend-main/src/app/shared/models/models.ts
export interface breadCrumbItems {
  label: string;
  active?: boolean;
}